requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
pydantic-settings==2.13.1
SQLAlchemy==2.0.49
alembic==1.18.4
asyncpg==0.31.0
GeoAlchemy2==0.18.4
shapely==2.1.2
redis==7.4.0
fastapi-cache2==0.2.2
Jinja2==3.1.6
python-dotenv==1.2.2
